third_party/WebKit/LayoutTests/editing/undo/undo-combined-delete-boundary-expected.txt

mac
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "rd "
PASS selection.focusOffset is 2
PASS selection.toString() is "word"
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected."
win
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "rd "
PASS selection.focusOffset is 2
PASS selection.toString() is "word"
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected."
unix
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "rd "
PASS selection.focusOffset is 2
PASS selection.toString() is "word"
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected."
android
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "rd "
PASS selection.focusOffset is 2
PASS selection.toString() is "word"
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected."
PASS successfullyParsed is true

TEST COMPLETE
